refactor: reduce test setup code
