+++ gateway/Cargo.toml
@@ -65,3 +65,4 @@ portpicker = { workspace = true }
 ring = { workspace = true }
 snailquote = "0.3.1"
 tempfile = { workspace = true }
+test-context = "0.1.4"
